[filterobjects]: added enums, updated files and tests to use enums, refactored code into functions, moved functions into library
