+++ tests/frontend/modules/test_filterobjects.py
@@ -20,6 +20,7 @@ import cellprofiler_core.pipeline
 import cellprofiler_core.preferences
 import cellprofiler_core.workspace
 import tests.frontend.modules
+from cellprofiler_library.opts.filterobjects import FilterMethod, FilterMode, OverlapAssignment
 
 INPUT_IMAGE = "input_image"
 INPUT_OBJECTS = "input_objects"
@@ -85,7 +86,7 @@ def make_classifier(
     os.write(fd, s)
     os.close(fd)
 
-    module.mode.value = cellprofiler.modules.filterobjects.MODE_CLASSIFIERS
+    module.mode.value = FilterMode.CLASSIFIERS.value
     module.rules_class.value = rules_class
     module.rules_directory.set_custom_path(os.path.dirname(filename))
     module.rules_file_name.value = os.path.split(filename)[1]
@@ -102,7 +103,7 @@ def test_zeros_single():
     module.x_name.value = INPUT_OBJECTS
     module.y_name.value = OUTPUT_OBJECTS
     module.measurements[0].measurement.value = TEST_FTR
-    module.filter_choice.value = cellprofiler.modules.filterobjects.FI_MAXIMAL
+    module.filter_choice.value = FilterMethod.MAXIMAL.value
     m = workspace.measurements
     m.add_measurement(INPUT_OBJECTS, TEST_FTR, numpy.zeros((0,)))
     module.run(workspace)
@@ -123,7 +124,7 @@ def test_zeros_per_object():
     module.enclosing_object_name.value = ENCLOSING_OBJECTS
     module.measurements[0].measurement.value = TEST_FTR
     module.filter_choice.value = (
-        cellprofiler.modules.filterobjects.FI_MAXIMAL_PER_OBJECT
+        FilterMethod.MAXIMAL_PER_OBJECT.value
     )
     m = workspace.measurements
     m.add_measurement(INPUT_OBJECTS, TEST_FTR, numpy.zeros((0,)))
@@ -138,7 +139,7 @@ def test_zeros_filter():
     module.x_name.value = INPUT_OBJECTS
     module.y_name.value = OUTPUT_OBJECTS
     module.measurements[0].measurement.value = TEST_FTR
-    module.filter_choice.value = cellprofiler.modules.filterobjects.FI_LIMITS
+    module.filter_choice.value = FilterMethod.LIMITS.value
     module.measurements[0].min_limit.value = 0
     module.measurements[0].max_limit.value = 1000
     m = workspace.measurements
@@ -160,7 +161,7 @@ def test_keep_single_min():
     module.x_name.value = INPUT_OBJECTS
     module.y_name.value = OUTPUT_OBJECTS
     module.measurements[0].measurement.value = TEST_FTR
-    module.filter_choice.value = cellprofiler.modules.filterobjects.FI_MINIMAL
+    module.filter_choice.value = FilterMethod.MINIMAL.value
     m = workspace.measurements
     m.add_measurement(INPUT_OBJECTS, TEST_FTR, numpy.array([2, 1]))
     module.run(workspace)
@@ -196,7 +197,7 @@ def test_keep_single_max():
     module.x_name.value = INPUT_OBJECTS
     module.y_name.value = OUTPUT_OBJECTS
     module.measurements[0].measurement.value = TEST_FTR
-    module.filter_choice.value = cellprofiler.modules.filterobjects.FI_MAXIMAL
+    module.filter_choice.value = FilterMethod.MAXIMAL.value
     m = workspace.measurements
     m.add_measurement(INPUT_OBJECTS, TEST_FTR, numpy.array([1, 2]))
     module.run(workspace)
@@ -222,7 +223,7 @@ def test_keep_one_min():
     module.enclosing_object_name.value = ENCLOSING_OBJECTS
     module.measurements[0].measurement.value = TEST_FTR
     module.filter_choice.value = (
-        cellprofiler.modules.filterobjects.FI_MINIMAL_PER_OBJECT
+        FilterMethod.MINIMAL_PER_OBJECT.value
     )
     m = workspace.measurements
     m.add_measurement(INPUT_OBJECTS, TEST_FTR, numpy.array([2, 1, 3, 4]))
@@ -249,7 +250,7 @@ def test_keep_one_max():
     module.enclosing_object_name.value = ENCLOSING_OBJECTS
     module.measurements[0].measurement.value = TEST_FTR
     module.filter_choice.value = (
-        cellprofiler.modules.filterobjects.FI_MAXIMAL_PER_OBJECT
+        FilterMethod.MAXIMAL_PER_OBJECT.value
     )
     m = workspace.measurements
     m.add_measurement(INPUT_OBJECTS, TEST_FTR, numpy.array([1, 2, 4, 3]))
@@ -275,10 +276,10 @@ def test_keep_maximal_most_overlap():
     module.enclosing_object_name.value = ENCLOSING_OBJECTS
     module.measurements[0].measurement.value = TEST_FTR
     module.filter_choice.value = (
-        cellprofiler.modules.filterobjects.FI_MAXIMAL_PER_OBJECT
+        FilterMethod.MAXIMAL_PER_OBJECT.value
     )
     module.per_object_assignment.value = (
-        cellprofiler.modules.filterobjects.PO_PARENT_WITH_MOST_OVERLAP
+        OverlapAssignment.PARENT_WITH_MOST_OVERLAP.value
     )
     m = workspace.measurements
     m.add_measurement(INPUT_OBJECTS, TEST_FTR, numpy.array([1, 4, 2]))
@@ -304,10 +305,10 @@ def test_keep_minimal_most_overlap():
     module.enclosing_object_name.value = ENCLOSING_OBJECTS
     module.measurements[0].measurement.value = TEST_FTR
     module.filter_choice.value = (
-        cellprofiler.modules.filterobjects.FI_MINIMAL_PER_OBJECT
+        FilterMethod.MINIMAL_PER_OBJECT.value
     )
     module.per_object_assignment.value = (
-        cellprofiler.modules.filterobjects.PO_PARENT_WITH_MOST_OVERLAP
+        OverlapAssignment.PARENT_WITH_MOST_OVERLAP.value
     )
     m = workspace.measurements
     m.add_measurement(INPUT_OBJECTS, TEST_FTR, numpy.array([4, 2, 3]))
@@ -336,7 +337,7 @@ def test_filter():
     module.x_name.value = INPUT_OBJECTS
     module.y_name.value = OUTPUT_OBJECTS
     module.measurements[0].measurement.value = TEST_FTR
-    module.filter_choice.value = cellprofiler.modules.filterobjects.FI_LIMITS
+    module.filter_choice.value = FilterMethod.LIMITS.value
     module.measurements[0].wants_minimum.value = True
     module.measurements[0].min_limit.value = my_min
     module.measurements[0].wants_maximum.value = True
@@ -367,7 +368,7 @@ def test_filter_min():
     module.x_name.value = INPUT_OBJECTS
     module.y_name.value = OUTPUT_OBJECTS
     module.measurements[0].measurement.value = TEST_FTR
-    module.filter_choice.value = cellprofiler.modules.filterobjects.FI_LIMITS
+    module.filter_choice.value = FilterMethod.LIMITS.value
     module.measurements[0].min_limit.value = my_min
     module.measurements[0].max_limit.value = 0.7
     module.measurements[0].wants_maximum.value = False
@@ -397,7 +398,7 @@ def test_filter_max():
     module.x_name.value = INPUT_OBJECTS
     module.y_name.value = OUTPUT_OBJECTS
     module.measurements[0].measurement.value = TEST_FTR
-    module.filter_choice.value = cellprofiler.modules.filterobjects.FI_LIMITS
+    module.filter_choice.value = FilterMethod.LIMITS.value
     module.measurements[0].min_limit.value = 0.3
     module.measurements[0].wants_minimum.value = False
     module.measurements[0].max_limit.value = my_max
@@ -432,7 +433,7 @@ def test_filter_two():
     for i in range(2):
         measurement_name = "measurement%d" % (i + 1)
         module.measurements[i].measurement.value = measurement_name
-        module.filter_choice.value = cellprofiler.modules.filterobjects.FI_LIMITS
+        module.filter_choice.value = FilterMethod.LIMITS.value
         module.measurements[i].min_limit.value = 0.3
         module.measurements[i].wants_minimum.value = False
         module.measurements[i].max_limit.value = my_max[i]
@@ -468,7 +469,7 @@ def test_renumber_other():
     module.x_name.value = INPUT_OBJECTS
     module.y_name.value = OUTPUT_OBJECTS
     module.measurements[0].measurement.value = TEST_FTR
-    module.filter_choice.value = cellprofiler.modules.filterobjects.FI_LIMITS
+    module.filter_choice.value = FilterMethod.LIMITS.value
     module.measurements[0].min_limit.value = my_min
     module.measurements[0].max_limit.value = my_max
     module.add_additional_object()
@@ -509,7 +510,7 @@ def test_renumber_other_object_numbers_reversed_does_not_relate():
     module.x_name.value = INPUT_OBJECTS
     module.y_name.value = OUTPUT_OBJECTS
     module.measurements[0].measurement.value = TEST_FTR
-    module.filter_choice.value = cellprofiler.modules.filterobjects.FI_LIMITS
+    module.filter_choice.value = FilterMethod.LIMITS.value
     module.measurements[0].min_limit.value = my_min
     module.measurements[0].max_limit.value = my_max
     module.add_additional_object()
@@ -629,7 +630,7 @@ def test_additional_objects_filter_correctly(get_images_for_issue_4509):
     )
     module.x_name.value = INPUT_OBJECTS
     module.y_name.value = OUTPUT_OBJECTS
-    module.filter_choice.value = cellprofiler.modules.filterobjects.FI_LIMITS
+    module.filter_choice.value = FilterMethod.LIMITS.value
     module.measurements[0].measurement.value = "Number_Object_Number"
     module.measurements[0].min_limit.value = 0.5
     module.measurements[0].max_limit.value = 2.5
@@ -676,7 +677,7 @@ def test_additional_objects_filter_correctly_2(get_images_for_issue_4509):
     )
     module.x_name.value = INPUT_OBJECTS
     module.y_name.value = OUTPUT_OBJECTS
-    module.filter_choice.value = cellprofiler.modules.filterobjects.FI_LIMITS
+    module.filter_choice.value = FilterMethod.LIMITS.value
     module.measurements[0].measurement.value = "Number_Object_Number"
     module.measurements[0].min_limit.value = 1.5
     module.measurements[0].max_limit.value = 2.5
@@ -713,14 +714,14 @@ def test_load_v3():
     assert isinstance(module, cellprofiler.modules.filterobjects.FilterObjects)
     assert module.x_name == "Things"
     assert module.y_name == "FilteredThings"
-    assert module.mode == cellprofiler.modules.filterobjects.MODE_MEASUREMENTS
+    assert module.mode == FilterMode.MEASUREMENTS.value
     assert (
         module.rules_directory.dir_choice
         == cellprofiler_core.preferences.DEFAULT_OUTPUT_FOLDER_NAME
     )
     assert module.rules_file_name == "myrules.txt"
     assert module.measurements[0].measurement == "Intensity_MeanIntensity_DNA"
-    assert module.filter_choice == cellprofiler.modules.filterobjects.FI_MINIMAL
+    assert module.filter_choice == FilterMethod.MINIMAL.value
 
 
 def test_load_v4():
@@ -740,8 +741,8 @@ def test_load_v4():
     assert isinstance(module, cellprofiler.modules.filterobjects.FilterObjects)
     assert module.y_name == "MyFilteredObjects"
     assert module.x_name == "MyObjects"
-    assert module.mode == cellprofiler.modules.filterobjects.MODE_MEASUREMENTS
-    assert module.filter_choice == cellprofiler.modules.filterobjects.FI_LIMITS
+    assert module.mode == FilterMode.MEASUREMENTS.value
+    assert module.filter_choice == FilterMethod.LIMITS.value
     assert (
         module.rules_directory.dir_choice
         == cellprofiler_core.preferences.DEFAULT_INPUT_FOLDER_NAME
@@ -782,8 +783,8 @@ def test_load_v5():
     assert isinstance(module, cellprofiler.modules.filterobjects.FilterObjects)
     assert module.y_name == "MyFilteredObjects"
     assert module.x_name == "MyObjects"
-    assert module.mode == cellprofiler.modules.filterobjects.MODE_MEASUREMENTS
-    assert module.filter_choice == cellprofiler.modules.filterobjects.FI_LIMITS
+    assert module.mode == FilterMode.MEASUREMENTS.value
+    assert module.filter_choice == FilterMethod.LIMITS.value
     assert (
         module.rules_directory.dir_choice
         == cellprofiler_core.preferences.DEFAULT_INPUT_FOLDER_NAME
@@ -825,9 +826,9 @@ def test_load_v6():
     assert isinstance(module, cellprofiler.modules.filterobjects.FilterObjects)
     assert module.y_name == "MyFilteredObjects"
     assert module.x_name == "MyObjects"
-    assert module.mode == cellprofiler.modules.filterobjects.MODE_MEASUREMENTS
-    assert module.filter_choice == cellprofiler.modules.filterobjects.FI_LIMITS
-    assert module.per_object_assignment == cellprofiler.modules.filterobjects.PO_BOTH
+    assert module.mode == FilterMode.MEASUREMENTS.value
+    assert module.filter_choice == FilterMethod.LIMITS.value
+    assert module.per_object_assignment == OverlapAssignment.BOTH.value
     assert (
         module.rules_directory.dir_choice
         == cellprofiler_core.preferences.DEFAULT_INPUT_FOLDER_NAME
@@ -869,11 +870,11 @@ def test_load_v7():
     assert isinstance(module, cellprofiler.modules.filterobjects.FilterObjects)
     assert module.y_name == "MyFilteredObjects"
     assert module.x_name == "MyObjects"
-    assert module.mode == cellprofiler.modules.filterobjects.MODE_MEASUREMENTS
-    assert module.filter_choice == cellprofiler.modules.filterobjects.FI_LIMITS
+    assert module.mode == FilterMode.MEASUREMENTS.value
+    assert module.filter_choice == FilterMethod.LIMITS.value
     assert (
         module.per_object_assignment
-        == cellprofiler.modules.filterobjects.PO_PARENT_WITH_MOST_OVERLAP
+        == OverlapAssignment.PARENT_WITH_MOST_OVERLAP.value
     )
     assert (
         module.rules_directory.dir_choice
@@ -918,7 +919,7 @@ def test_filter_by_rule():
         fd.close()
         rules_dir, rules_file = os.path.split(rules_path)
         module.x_name.value = "MyObjects"
-        module.mode.value = cellprofiler.modules.filterobjects.MODE_RULES
+        module.mode.value = FilterMode.RULES.value
         module.rules_file_name.value = rules_file
         module.rules_directory.dir_choice = (
             cellprofiler_core.preferences.ABSOLUTE_FOLDER_NAME
@@ -959,7 +960,7 @@ def test_filter_by_3_class_rule():
             )
             rules_dir, rules_file = os.path.split(rules_path)
             module.x_name.value = "MyObjects"
-            module.mode.value = cellprofiler.modules.filterobjects.MODE_RULES
+            module.mode.value = FilterMode.RULES.value
             module.rules_file_name.value = rules_file
             module.rules_directory.dir_choice = (
                 cellprofiler_core.preferences.ABSOLUTE_FOLDER_NAME
@@ -995,7 +996,7 @@ def test_filter_by_fuzzy_rule():
         fd.close()
         rules_dir, rules_file = os.path.split(rules_path)
         module.x_name.value = "MyObjects"
-        module.mode.value = cellprofiler.modules.filterobjects.MODE_RULES
+        module.mode.value = FilterMode.RULES.value
         module.rules_file_name.value = rules_file
         module.rules_directory.dir_choice = (
             cellprofiler_core.preferences.ABSOLUTE_FOLDER_NAME
@@ -1027,7 +1028,7 @@ def test_discard_border_objects():
     workspace, module = make_workspace({INPUT_OBJECTS: labels})
     module.x_name.value = INPUT_OBJECTS
     module.y_name.value = OUTPUT_OBJECTS
-    module.mode.value = cellprofiler.modules.filterobjects.MODE_BORDER
+    module.mode.value = FilterMode.BORDER.value
     module.run(workspace)
     output_objects = workspace.object_set.get_objects(OUTPUT_OBJECTS)
     assert numpy.all(expected == output_objects.segmented)
@@ -1055,7 +1056,7 @@ def test_discard_mask_objects():
 
     module.x_name.value = INPUT_OBJECTS
     module.y_name.value = OUTPUT_OBJECTS
-    module.mode.value = cellprofiler.modules.filterobjects.MODE_BORDER
+    module.mode.value = FilterMode.BORDER.value
     module.run(workspace)
     output_objects = workspace.object_set.get_objects(OUTPUT_OBJECTS)
     assert numpy.all(expected == output_objects.segmented)
@@ -1080,9 +1081,9 @@ def test_unedited_segmented():
     input_objects.unedited_segmented = unedited
     module.x_name.value = INPUT_OBJECTS
     module.y_name.value = OUTPUT_OBJECTS
-    module.mode.value = cellprofiler.modules.filterobjects.MODE_MEASUREMENTS
+    module.mode.value = FilterMode.MEASUREMENTS.value
     module.measurements[0].measurement.value = TEST_FTR
-    module.filter_choice.value = cellprofiler.modules.filterobjects.FI_MINIMAL
+    module.filter_choice.value = FilterMethod.MINIMAL.value
     m = workspace.measurements
     m[INPUT_OBJECTS, TEST_FTR] = numpy.array([2, 1])
     module.run(workspace)
@@ -1111,9 +1112,9 @@ def test_small_removed_segmented():
     workspace.object_set.add_objects(input_objects, INPUT_OBJECTS)
     module.x_name.value = INPUT_OBJECTS
     module.y_name.value = OUTPUT_OBJECTS
-    module.mode.value = cellprofiler.modules.filterobjects.MODE_MEASUREMENTS
+    module.mode.value = FilterMode.MEASUREMENTS.value
     module.measurements[0].measurement.value = TEST_FTR
-    module.filter_choice.value = cellprofiler.modules.filterobjects.FI_MINIMAL
+    module.filter_choice.value = FilterMethod.MINIMAL.value
     m = workspace.measurements
     m[INPUT_OBJECTS, TEST_FTR] = numpy.array([2, 1])
     module.run(workspace)
@@ -1242,7 +1243,7 @@ def test_measurements():
 
     module.measurements[0].measurement.value = TEST_FTR
 
-    module.filter_choice.value = cellprofiler.modules.filterobjects.FI_MAXIMAL
+    module.filter_choice.value = FilterMethod.MAXIMAL.value
 
     measurements = workspace.measurements
 
@@ -1296,7 +1297,7 @@ def test_discard_border_objects_volume():
 
     module.x_name.value = INPUT_OBJECTS
     module.y_name.value = OUTPUT_OBJECTS
-    module.mode.value = cellprofiler.modules.filterobjects.MODE_BORDER
+    module.mode.value = FilterMode.BORDER.value
 
     module.run(workspace)
 
@@ -1326,10 +1327,10 @@ def test_keep_maximal_per_object_most_overlap_volume():
     module.enclosing_object_name.value = ENCLOSING_OBJECTS
     module.measurements[0].measurement.value = TEST_FTR
     module.filter_choice.value = (
-        cellprofiler.modules.filterobjects.FI_MAXIMAL_PER_OBJECT
+        FilterMethod.MAXIMAL_PER_OBJECT.value
     )
     module.per_object_assignment.value = (
-        cellprofiler.modules.filterobjects.PO_PARENT_WITH_MOST_OVERLAP
+        OverlapAssignment.PARENT_WITH_MOST_OVERLAP.value
     )
 
     m = workspace.measurements
@@ -1366,9 +1367,9 @@ def test_keep_minimal_per_object_both_parents_image():
     module.enclosing_object_name.value = ENCLOSING_OBJECTS
     module.measurements[0].measurement.value = TEST_FTR
     module.filter_choice.value = (
-        cellprofiler.modules.filterobjects.FI_MINIMAL_PER_OBJECT
+        FilterMethod.MINIMAL_PER_OBJECT.value
     )
-    module.per_object_assignment.value = cellprofiler.modules.filterobjects.PO_BOTH
+    module.per_object_assignment.value = OverlapAssignment.BOTH.value
 
     m = workspace.measurements
 
@@ -1403,9 +1404,9 @@ def test_keep_maximal_both_parents_volume():
     module.enclosing_object_name.value = ENCLOSING_OBJECTS
     module.measurements[0].measurement.value = TEST_FTR
     module.filter_choice.value = (
-        cellprofiler.modules.filterobjects.FI_MAXIMAL_PER_OBJECT
+        FilterMethod.MAXIMAL_PER_OBJECT.value
     )
-    module.per_object_assignment.value = cellprofiler.modules.filterobjects.PO_BOTH
+    module.per_object_assignment.value = OverlapAssignment.BOTH.value
 
     m = workspace.measurements
 
@@ -1432,7 +1433,7 @@ def test_keep_maximal_volume():
     module.x_name.value = INPUT_OBJECTS
     module.y_name.value = OUTPUT_OBJECTS
     module.measurements[0].measurement.value = TEST_FTR
-    module.filter_choice.value = cellprofiler.modules.filterobjects.FI_MAXIMAL
+    module.filter_choice.value = FilterMethod.MAXIMAL.value
 
     m = workspace.measurements
 
@@ -1466,7 +1467,7 @@ def test_keep_removed_volume():
     module.keep_removed_objects.value = True
     module.removed_objects_name.value = "RemovedObjects"
     module.measurements[0].measurement.value = TEST_FTR
-    module.filter_choice.value = cellprofiler.modules.filterobjects.FI_MAXIMAL
+    module.filter_choice.value = FilterMethod.MAXIMAL.value
 
     m = workspace.measurements
